pyproject.toml
[tool.poetry]
name = "melody.web"
version = "1.0.0"
description = "All your music, in one place."
authors = ["MelodyKit"]
license = "MIT"

readme = "README.md"

homepage = "https://melodykit.app/"
repository = "https://github.com/MelodyKit/melody.web"

keywords = ["python", "api", "kit", "web"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
    "Typing :: Typed",
]

[tool.poetry.urls]
Chat = "https://melodykit.app/discord"
Issues = "https://github.com/MelodyKit/melody.web/issues"

[[tool.poetry.packages]]
include = "melody"

[tool.poetry.scripts]
"melody.bot" = "melody.bot.main:bot"
"melody.kit" = "melody.kit.main:run"
"melody.web" = "melody.web.main:web"

[tool.poetry.dependencies]
# melody
python = ">= 3.8"
versions = "^2.0.0"
attrs = "^23.2.0"
loguru = "^0.7.2"
# melody.kit
qrcode = "^7.4.2"
authlib = "^1.3.0"
httpx = "^0.26.0"  # for `authlib`
itsdangerous = "^2.1.2"  # for `starlette`
fastapi = "^0.109.2"
toml = "^0.10.2"
pendulum = "^3.0.0"
edgedb = "^1.8.0"
iters = "^0.15.0"
redis = "^5.0.1"
argon2-cffi = "^23.1.0"
click = "^8.1.7"
entrypoint = "^2.0.0"
email-validator = "^2.1.0"
colors-core = "^1.0.1"
typing-extensions = "^4.9.0"
async-extensions = "^1.4.1"
python-multipart = "^0.0.9"
pyotp = "^2.9.0"
# melody.web
jinja2 = "^3.1.3"
# melody.shared
aiohttp = "^3.9.1"
aiofiles = "^23.2.1"
cattrs = "^23.2.3"
typing-aliases = "^1.4.1"
yarl = "^1.9.4"
pillow = "^10.2.0"

[tool.poetry.dependencies.aiosmtplib]  # melody.kit
version = "^3.0.1"
python = "^3.8"

[tool.poetry.dependencies.uvicorn]  # melody.kit
version = "^0.27.0"
extras = ["standard"]

[tool.poetry.dependencies."discord.py"]  # melody.bot
version = "^2.3.2"

[tool.poetry.group.format.dependencies]
ruff = "0.2.2"

[tool.poetry.group.check.dependencies]
mypy = "1.8.0"

types-toml = "0.10.8.7"
types-redis = "4.6.0.20240218"
types-aiofiles = "23.2.0.20240106"
types-qrcode = "7.4.0.20240106"
types-pillow = "10.2.0.20240213"

[tool.poetry.group.check.dependencies.pre-commit]
version = "3.6.1"
python = ">= 3.9"

[tool.poetry.group.test.dependencies]
coverage = "7.4.1"
pytest = "8.0.1"
pytest-cov = "4.1.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "1.5.3"
mkdocs-material = "9.5.9"

[tool.poetry.group.docs.dependencies.mkdocstrings]
version = "0.24.0"
extras = ["python"]

[tool.poetry.group.release]
optional = true

[tool.poetry.group.release.dependencies]
changelogging = "1.3.0"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
ignore = [
    "E402",  # module level import not at top of file
    "E741",  # ambiguous variable name
]

[tool.mypy]
strict = true

[tool.changelogging]
name = "melody.web"
version = "1.0.0"
url = "https://github.com/MelodyKit/melody.web"
directory = "changes"
output = "CHANGELOG.md"

start_string = "<!-- changelogging: start -->"

title_format = "{version} ({date})"
issue_format = "[#{issue}]({url}/pull/{issue})"

bullet = "-"
wrap = true
wrap_size = 100

display = ["feature", "change", "fix", "security", "deprecation", "removal", "internal"]

[build-system]
requires = ["poetry-core >= 1.8.1"]
build-backend = "poetry.core.masonry.api"
